Rename deep-getter.ts to column-prop-getters.ts
